Organize event modules into package
