+++ bang_py/events/__init__.py
@@ -1,0 +1,7 @@
+"""Event deck implementations and logic for optional expansions.
+
+This package contains helpers for constructing event decks, applying their
+effects during gameplay, and notifying game components of event-driven
+changes. It centralizes all event-related utilities used by the Bang game
+manager.
+"""
